MNT: rearrange procedure dataclasses to atef.procedures
